src/server/requirements.txt
bcrypt==4.0.1
fastapi==0.110.0
mysql-connector-python==8.3.0
passlib[bcrypt]==1.7.4
pycryptodome==3.20.0
python-jose[cryptography]==3.3.0
pydantic==2.6.3
python-multipart==0.0.9
uvicorn==0.28.0
